chore: Post, Comment에 findByMemberId 추가 (#109)
* fix: 충돌 해결

* feat: PostService 작업

* feat: create.html 작업

* feat: list.html 작업

* feat: PostController 추가

* refactor: Community, Post 전면 재 수정

* fix: 게시글 작성시 발생하는 403에러 해결

* feat: 게시글 삭제 기능 구현

* chore: 에러 확인용

* fix: 게시글 작성 안되는 에러 해결

* feat: 게시글 수정 기능 추가

* feat: InitData로 초기 게시글 작성

* feat: 게시글 글번호가 큰 것부터 내림차순으로 정렬

* refactor: Rq, RsData 도입

* refactor: community 코드 수정

* chore: 매핑 중복 수정

* feat: 게시판 페이징 기능 추가

* fix: 페이지 끝에서도 '>'작동하는 오류 해결

* fix: list에서 작성자 보이게끔 수정

* chore: list에서 작성일자 제거

* feat: 댓글 작성 기능 구현

* chore: 외래키 관련 에러 발생

* feat: 댓글 삭제 기능 구현

* feat: 댓글 수정 기능 구현

* chore: 작성자 nickname으로 출력

* feat: 조회수 증가 기능 구현

* feat: 다중 게시판기능 일부 구현

* chore: 게시판 선택 위치 조절

* fix: 게시글 수정을 취소하면 발생하던 ã 에러 수정

* feat: 자신이 작성한 게시글, 댓글만 수정 삭제가 가능하도록 구현

* refactor: ADMIN은 모든 댓글 삭제권한 가지도록 설정

* chore: Post, Comment에 findByMemberId 추가

+++ src/main/java/com/twenty/inhub/boundedContext/post/service/PostService.java
@@ -4,6 +4,7 @@ import com.amazonaws.services.kms.model.NotFoundException;
 import com.twenty.inhub.boundedContext.comment.entity.Comment;
 import com.twenty.inhub.boundedContext.comment.repository.CommentRepository;
 import com.twenty.inhub.boundedContext.member.entity.Member;
+import com.twenty.inhub.boundedContext.member.entity.MemberRole;
 import com.twenty.inhub.boundedContext.member.repository.MemberRepository;
 import com.twenty.inhub.boundedContext.post.dto.PostDto;
 import com.twenty.inhub.boundedContext.post.entity.Post;
@@ -47,12 +48,15 @@ public class PostService {
         return postDtoList;
     }
 
-    public void updatePost(PostDto postDto) {
+    public void updatePost(PostDto postDto, Member member) {
         Post post = postRepository.findById(postDto.getId()).orElse(null);
-        if (post != null) {
+        if (post != null && post.isCreatedBy(member)) {
             post.setTitle(postDto.getTitle());
             post.setContent(postDto.getContent());
             postRepository.save(post);
+        }
+        else {
+            throw new IllegalArgumentException("이 게시글을 수정할 권한이 없습니다.");
         }
     }
 
@@ -60,21 +64,25 @@ public class PostService {
         return postRepository.findById(id).orElse(null);
     }
 
-    public void deletePost(Long id) {
+    public void deletePost(Long id, Member member) {
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("게시글을 찾을 수 없습니다."));
 
-        // 댓글 삭제
-        List<Comment> comments = post.getComments();
-        commentRepository.deleteAll(comments);
+        if (post.isCreatedBy(member) || member.getRole() == MemberRole.ADMIN) {
+            // 댓글 삭제
+            List<Comment> comments = post.getComments();
+            commentRepository.deleteAll(comments);
 
-        // 회원 엔티티에서 게시글 제거
-        Member member = post.getMember();
-        member.getPosts().remove(post);
-        memberRepository.save(member);
+            // 회원 엔티티에서 게시글 제거
+            member.getPosts().remove(post);
+            memberRepository.save(member);
 
-        // 게시글 삭제
-        postRepository.deleteById(id);
+            // 게시글 삭제
+            postRepository.delete(post);
+        }
+        else {
+            throw new IllegalArgumentException("이 게시글을 삭제할 권한이 없습니다.");
+        }
     }
 
     public Page<Post> getList(String board, int page) {
@@ -107,5 +115,9 @@ public class PostService {
             }
         }
         return post;
+    }
+
+    public List<Post> findByMemberId(Long memberId) {
+        return postRepository.findByMemberId(memberId);
     }
 }
